[Refactor](공통): Change 패키지 폴더 구조 재구성
