+++ src/main/java/com/test/main/stock/orderlist/OrderList.java
@@ -1,4 +1,4 @@
-package com.test.main.stock;
+package com.test.main.stock.orderlist;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -8,13 +8,13 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/stock/orderlist.do")
+@WebServlet("/stock/order/orderlist.do")
 public class OrderList extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/views/stock/orderlist.jsp");
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/views/stock/order/orderlist.jsp");
         dispatcher.forward(req, resp);
     }
 }
